refactor: remove Circular dependency
