+++ packages/vtable/src/scenegraph/layout/update-width.ts
@@ -4,20 +4,19 @@ import { CartesianAxis } from '../../components/axis/axis';
 import { getStyleTheme } from '../../core/tableHelper';
 import type { BaseTableAPI } from '../../ts-types/base-table';
 import type { IProgressbarColumnBodyDefine } from '../../ts-types/list-table/define/progressbar-define';
-import { dealWithCustom, getCustomCellMergeCustom } from '../component/custom';
+import { dealWithCustom } from '../component/custom';
 import type { Group } from '../graphic/group';
-import type { Icon } from '../graphic/icon';
 import { updateImageCellContentWhileResize } from '../group-creater/cell-type/image-cell';
 import { createProgressBarCell } from '../group-creater/cell-type/progress-bar-cell';
 import { createSparkLineCellGroup } from '../group-creater/cell-type/spark-line-cell';
-import { resizeCellGroup } from '../group-creater/column-helper';
+import { resizeCellGroup, getCustomCellMergeCustom } from '../group-creater/cell-helper';
 import type { Scenegraph } from '../scenegraph';
 import { getCellMergeInfo } from '../utils/get-cell-merge';
 import { getProp } from '../utils/get-prop';
 import { isMergeCellGroup } from '../utils/is-merge-cell-group';
 import { getQuadProps } from '../utils/padding';
 import { updateCellContentWidth } from '../utils/text-icon-layout';
-import { computeRowHeight, computeRowsHeight } from './compute-row-height';
+import { computeRowHeight } from './compute-row-height';
 import { updateCellHeightForRow } from './update-height';
 import { getHierarchyOffset } from '../utils/get-hierarchy-offset';
 import { getCellMergeRange } from '../../tools/merge-range';
